Extract phonological inventory display into its own view and rework PhonologyView to support multiple displays
