+++ langforge-app-phonology/src/main/resources/fxml/PhonologicalInventoryView.fxml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.Double?>
+<?import java.lang.String?>
+
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<?import javafx.collections.FXCollections?>
+
+<fx:root type="net.calebscode.langforge.app.phonology.view.PhonologicalInventoryView" xmlns:fx="http://javafx.com/fxml">
+	<VBox style="-fx-padding: 5px" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
+		<Label text="Add Phonemes" />
+		<HBox VBox.vgrow="NEVER" spacing="5">
+			<Button text="Add IPA Consonants" onMouseClicked="#showIpaConsonantPicker" HBox.hgrow="ALWAYS">
+				<maxWidth><Double fx:constant="MAX_VALUE" /></maxWidth>
+			</Button>
+			<Button text="Add IPA Vowels" onMouseClicked="#showIpaVowelPicker" HBox.hgrow="ALWAYS">
+				<maxWidth><Double fx:constant="MAX_VALUE" /></maxWidth>
+			</Button>
+		</HBox>
+		<Label text="Phonological Inventory" />
+		<TableView fx:id="phonemesTable" VBox.vgrow="ALWAYS">
+			<columnResizePolicy><TableView fx:constant="CONSTRAINED_RESIZE_POLICY_NEXT_COLUMN"/></columnResizePolicy>
+		</TableView>
+	</VBox>
+</fx:root>
+
