+++ langforge-app-phonology/src/main/resources/fxml/PhonologyView.fxml
@@ -1,28 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.Double?>
+<?import java.lang.String?>
 
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.SplitPane?>
 
 <?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 
+<?import javafx.collections.FXCollections?>
+
 <fx:root type="net.calebscode.langforge.app.phonology.view.PhonologyView" xmlns:fx="http://javafx.com/fxml">
-	<VBox AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
-		<HBox VBox.vgrow="NEVER">		
-			<Button text="Add IPA Consonants" onMouseClicked="#showIpaConsonantPicker" HBox.hgrow="ALWAYS">
-				<maxWidth><Double fx:constant="MAX_VALUE" /></maxWidth>
-			</Button>
-			<Button text="Add IPA Vowels" onMouseClicked="#showIpaVowelPicker" HBox.hgrow="ALWAYS">
-				<maxWidth><Double fx:constant="MAX_VALUE" /></maxWidth>
-			</Button>
-		</HBox>
-		<TableView fx:id="phonemesTable" VBox.vgrow="ALWAYS">
-			<columnResizePolicy><TableView fx:constant="CONSTRAINED_RESIZE_POLICY_NEXT_COLUMN"/></columnResizePolicy>
-		</TableView>
-	</VBox>
+	<SplitPane AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
+		<ListView fx:id="viewSelector" minWidth="150" maxWidth="150">
+			<items>
+				<FXCollections fx:factory="observableArrayList">
+					<String fx:value="Phonological Inventory" />
+					<String fx:value="Syllable Patterns" />
+				</FXCollections>
+			</items>
+		</ListView>
+		<VBox fx:id="contentContainer">
+		</VBox>
+	</SplitPane>
 </fx:root>
